financial_kg/requirements.txt
google-generativeai
langchain
langchain-google-genai
neo4j
pandas
numpy
python-dotenv
pydantic
aiohttp
asyncio
requests
beautifulsoup4
lxml
datasets
huggingface-hub
sentence-transformers
scikit-learn
plotly
matplotlib
seaborn
jupyter
ipykernel
python-dateutil
pytz
tqdm
rich
